backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
PyJWT
cryptography
requests
qrcode
Pillow
python-dotenv